% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maptools.R
\name{addCircles_map}
\alias{addCircles_map}
\title{add circles to an existig map (proxy argument)}
\usage{
addCircles_map(
  proxy,
  data,
  points_col,
  pal,
  radius = rlang::expr(10),
  label_formula = rlang::expr(~paste0("Person ID: ", as.character(data[[points_col]])))
)
}
\arguments{
\item{proxy}{Existing Map}

\item{data}{Dataframe containing "lat" and "lng"}

\item{points_col}{Chr. Column which Identify different points}

\item{pal}{Colorpalette function}

\item{radius}{an expression (formula)}

\item{label_formula}{an expression for the label. By default is used the \code{points_col} column}
}
\value{
A Leaflet Map
}
\description{
add circles to an existig map (proxy argument)
}
