% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_ints.R
\name{all_ints}
\alias{all_ints}
\title{Are all elements of a vector integers?}
\usage{
all_ints(xs)
}
\arguments{
\item{xs}{A vector of any type.}
}
\value{
Return TRUE if all elements of a given vector are integers,
        FALSE otherwise.
}
\description{
Check if a vector has all integer elements.
}
\examples{
all_ints(iris[[4]])
all_ints(mtcars[[2]])
unique(mtcars[[2]])
str(mtcars[[2]])
}
