% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize the first letter.}
\usage{
capitalize(char_vec)
}
\arguments{
\item{char_vec}{A character vector (or a vector of strings).}
}
\value{
A vector of 1st-letter capitalized strings.
}
\description{
Capitalize the first letter of each element in a character vector.
NA elements are returned as NA.
}
\examples{
s = c("gmlang", NA, "hello", "Jan", "feb")
capitalize(s)
}
