% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkboxes-allon-alloff.R
\name{checkboxes_allon_alloff_server}
\alias{checkboxes_allon_alloff_server}
\title{Implement the control of checkboxes by all-on/all-off buttons.}
\usage{
checkboxes_allon_alloff_server(
  id,
  label = NULL,
  choices,
  init_checked = choices
)
}
\arguments{
\item{id}{String. The input slot that will be used to access the value.}

\item{label}{String. Display label/title for the checkboxes group, or
NULL (default) for no label.}

\item{choices}{Character vector. List of values to show checkboxes for.
If elements of the list are named then that name rather than the value
is displayed to the user. The values should be strings; other types
(such as logicals and numbers) will be coerced to strings.}

\item{init_checked}{Character vector. List of values checked initially when
displaying the group of checkboxes for the first time. Default is
to have all choices checked initally.}
}
\value{
A module server function that can be called in a reactive env.
}
\description{
The server component of the Shiny module for creating a group of pretty
checkboxes with an all-on and an all-off control buttons.
}
\examples{
inst/examples/ex-checkboxes-allon-alloff.R
}
\seealso{
\code{\link{checkboxes_allon_alloff_ui}} for the UI, and
         \code{\link{mk_checkboxes}} for how the group checkboxes are made.
}
