% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkboxes-allon-alloff.R
\name{checkboxes_allon_alloff_ui}
\alias{checkboxes_allon_alloff_ui}
\title{A group of checkboxes with all-on/all-off control buttons.}
\usage{
checkboxes_allon_alloff_ui('id', align_onoff_bttns = 'horizontal',
    hbttns_height = "100\%", hbttns_width = "100\%")
checkboxes_allon_alloff_ui('id', align_onoff_bttns = 'vertical',
    line_breaks_bw_cboxes_n_vbttns = 1, vbttns_height = "100\%",
    vbttns_width = "100\%")
}
\arguments{
\item{id}{String. The input slot that will be used to access the value.}

\item{align_onoff_bttns}{String. Possible values are 'vertical' (default) or
'horizontal'. If 'vertical', the all-on and all-off buttons are stacked
vertically. Otherwise, they are placed horizontally side by side.}

\item{line_breaks_bw_cboxes_n_vbttn1, line_breaks_bw_vbttn1_n_vbttn2}{Integer.
Number of line breaks between the checkboxes group and the 1st
vertical button, and between the 1st and 2nd vertically stacked
buttons. Only used when `align_onoff_bttns = 'vertical'`.}

\item{hbttns_height, hbttns_width}{The total amount of height and width
to use for the entire column of the horizontally aligned buttons group.
Their values will be passed into `fillRow()`. Only used when
`align_onoff_bttns = 'horizontal'`.}
}
\value{
Must use it together with its counterpart server.
}
\description{
The UI component of the Shiny module for creating a group of pretty
checkboxes with an all-on and an all-off control buttons.
}
\examples{
inst/examples/ex-checkboxes-allon-alloff.R
}
\seealso{
\code{\link{checkboxes_allon_alloff_server}} for the server.
}
