% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata_module.R
\name{csvFile}
\alias{csvFile}
\title{Server function of Load - data CSV}
\usage{
csvFile(input, output, session, stringsAsFactors = FALSE)
}
\arguments{
\item{stringsAsFactors}{FALSE by default.}
}
\value{
A reactive Dataframe
}
\description{
@seealso \link{csvFileInput}
}
