% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata_module.R
\name{csvFileInput}
\alias{csvFileInput}
\title{Ui function foa loading csv Data}
\usage{
csvFileInput(id, label = "CSV file")
}
\arguments{
\item{id}{Id module}
}
\value{
UI module
}
\description{
Ui function foa loading csv Data
}
