% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_datatransform.R
\name{dataTransform}
\alias{dataTransform}
\title{DataTransform server function}
\usage{
dataTransform(input, output, session)
}
\value{
A Server function
}
\description{
DataTransform server function
}
\examples{
require(shiny)
if(interactive()) {

ui <- fluidPage(
  dataTransformUi("prova"))

server <- function(input, output, session) {
  callModule(dataTransform, "prova")
}

shinyApp(ui, server )
}
}
