% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_datatransform.R
\name{dataTransformUi}
\alias{dataTransformUi}
\title{Data transformation module}
\usage{
dataTransformUi(id)
}
\arguments{
\item{id}{}
}
\value{
Ui Module
}
\description{
Ui module for simulating data and apply to them transformation.
}
