% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{df_hist}
\alias{df_hist}
\title{create an histogram object}
\usage{
df_hist(df, x_col, title = "Aggregate", custom = NULL)
}
\arguments{
\item{df}{df}

\item{x_col}{col to represent}

\item{title}{optional title}

\item{custom}{optional cusyom}
}
\value{
histogram object
}
\description{
create an histogram object
}
