% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncalls.R
\name{.modify_call}
\alias{.modify_call}
\title{Modify call of a function}
\usage{
.modify_call(x, param, val)
}
\arguments{
\item{x}{function call object}

\item{param}{chr. Named argument of the function.}

\item{val}{New val of the argument.}
}
\value{
a function call
}
\description{
Modify call of a function
}
