% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_baseplot_server}
\alias{download_baseplot_server}
\title{Implement the download button for a base plot.}
\usage{
download_baseplot_server(id, obj, fname = "download", dpi = 156, ...)
}
\arguments{
\item{id}{String. The input slot that will be used to access the value.}

\item{obj}{The ggplot2 object to be saved.}

\item{fname}{String. Name of the file under which to save the object.}

\item{dpi}{Number. The resolution of the saved png file. Default = 156.}

\item{...}{Arguments that can be passed into `grDevices::png()`. For
example, `units`, `width` and `height`.}
}
\value{
A module server function that can be called in a reactive env.
}
\description{
The server component of the Shiny module for creating a pretty download
button for downloading a base plot as a png file.
}
\examples{
inst/examples/ex-download.R
}
\seealso{
\code{\link{download_ui}} for the UI.
}
