% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_pdf_server}
\alias{download_pdf_server}
\title{Implement the download button for a pdf report.}
\usage{
download_pdf_server(id, template_fname, fname = "report", ...)
}
\arguments{
\item{id}{String. The input slot that will be used to access the value.}

\item{template_fname}{String. The file name of the Rmd template for rendering the report.}

\item{fname}{String. Name of the file under which to save the object.}

\item{...}{Arguments that can be passed into the `params` argument inside
`rmarkdown::render()`.}
}
\value{
A module server function that can be called in a reactive env.
}
\description{
The server component of the Shiny module for creating a pretty pdf download
button.
}
\examples{
inst/examples/ex-download.R
}
\seealso{
\code{\link{download_ui}} for the UI.
}
