% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_server}
\alias{download_server}
\title{Implement the download button.}
\usage{
download_server(id, obj, fname = "download", ftype = "csv", ...)
}
\arguments{
\item{id}{String. The input slot that will be used to access the value.}

\item{obj}{The object to be saved. Either a data frame or a ggplot2 object.}

\item{fname}{String. Name of the file under which to save the object.}

\item{ftype}{String. Type of the file to save. One of three values:
'csv' (default), 'png', 'jpg'.}

\item{...}{Arguments that can be passed into `ggplot2::ggsave()`. For
example `dpi` (image resolution), `width`, and `height`.}
}
\value{
A module server function that can be called in a reactive env.
}
\description{
The server component of the Shiny module for creating a pretty download
button for downloading a csv, png, or jpg file.
}
\seealso{
\code{\link{download_ui}} for the UI.
}
