% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_ui}
\alias{download_ui}
\title{Download button.}
\usage{
download_ui(id, label = "Download")
}
\arguments{
\item{id}{String. The input slot that will be used to access the value.}

\item{label}{String. Text (default = 'Download') to be shown on the download
button.}
}
\value{
A list of HTML elements that can be added to a UI definition and
        accessed by the server component.
}
\description{
The UI component of the Shiny module for creating a pretty download button
for downloading a csv, png, or jpg file.
}
\examples{
inst/examples/ex-download.R
}
\seealso{
\code{\link{download_server}} for the server.
}
