% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown-box.R
\name{dropdown_box_server}
\alias{dropdown_box_server}
\title{Implement a single dropdown box.}
\usage{
dropdown_box_server(id)
}
\arguments{
\item{id}{String. The input slot that will be used to access the value.}
}
\value{
A module server function that can be called in a reactive env.
}
\description{
The server component of the Shiny module for creating a single dropdown box.
}
\examples{
inst/examples/ex-dropdown-box.R
}
\seealso{
\code{\link{dropdown_box_ui}} for the UI.
}
