% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown-box.R
\name{dropdown_box_ui}
\alias{dropdown_box_ui}
\title{A plain single dropdown box.}
\usage{
dropdown_box_ui(id, label = NULL, choices)
}
\arguments{
\item{id}{String. The input slot that will be used to access the value.}

\item{label}{String. Display label/title for the dropdown box, or
NULL (default) for no label.}

\item{choices}{Character vector. List of values listed under the dropdown box
for user to select. If elements of the list are named then that name
rather than the value is displayed. The values should be strings;
other types (such as logicals and numbers) will be coerced to strings.}
}
\value{
Must use it together with its counterpart server.
}
\description{
The UI component of the Shiny module for creating a plain single dropdown
box.
}
\examples{
inst/examples/ex-dropdown-box.R
}
\seealso{
\code{\link{dropdown_box_server}} for the server
         \code{\link{mk_dropdown_box}} for how to make a plain dropdown box.
}
