% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_types}
\alias{extract_types}
\title{Extract class of variables from a dataframe}
\usage{
extract_types(x, num_to_chr = "id")
}
\arguments{
\item{x}{Dataframe}

\item{num_to_chr}{Character string. Numerical Columns ending/starting with \code{num_to_chc}
will be considered character.}
}
\value{
list of numerical, character, datevars
}
\description{
Extract class of variables from a dataframe. Factors are considered character
}
