% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_types_original}
\alias{extract_types_original}
\title{Extract class of variables from a dataframe dataframe}
\usage{
extract_types_original(x)
}
\arguments{
\item{x}{df}
}
\value{
list
}
\description{
Extract class of variables from a dataframe dataframe
}
