% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata_module.R
\name{featherFile}
\alias{featherFile}
\title{Server function of  load Feather data module}
\usage{
featherFile(input, output, session)
}
\value{
A Reactive Dataframe
}
\description{
Server function of  load Feather data module
}
