% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata_module.R
\name{featherFileInput}
\alias{featherFileInput}
\title{Load Feather file in shiny}
\usage{
featherFileInput(id, label = "FEATHER format")
}
\arguments{
\item{id}{}
}
\value{
A UI module
}
\description{
Load Feather file in shiny
}
