% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterpanel.R
\name{filterpanel}
\alias{filterpanel}
\title{Filter Shiny Module}
\usage{
filterpanel(input, output, session, data)
}
\arguments{
\item{data}{A Dataframe to filter. Accept numerical and chr vars}
}
\value{
A reactive Dataframe
}
\description{
Filter Shiny Module
}
\examples{
require(shiny)
if(interactive()) {

ui <- fluidPage(

  filterpanelUi("prova"),

  mainPanel(tableOutput("tbl"))

)

server <- function(input, output, session) {

  dt <- callModule(filterpanel, "prova", data = CO2)

  output$tbl <- renderTable({

    dt()
  })

}
shinyApp(ui, server)

}
}
