% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlize.R
\name{htmlize_spaces}
\alias{htmlize_spaces}
\title{Multiple Spaces}
\usage{
htmlize_spaces(n)
}
\arguments{
\item{n}{Integer. The number of spaces to make.}
}
\value{
HTML code of n spaces.
}
\description{
Creates n spaces.
}
\examples{
htmlize_spaces(5)
}
