% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlize.R
\name{htmlize_title}
\alias{htmlize_title}
\title{Add html styling to page or section title.}
\usage{
htmlize_title(title, type = "page")
}
\arguments{
\item{title}{String. Title of a page or a section.}

\item{type}{String. Either 'page' (default) or 'section'.}
}
\value{
HTML code wrapped around the input title.
}
\description{
Make the title bigger, emphasized and blue colored.
}
\examples{
htmlize_title("This app is designed for Line Optimization", type = 'page')
htmlize_title("Simulate Tab", type = 'section')
}
