% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlize.R
\name{htmlize_uoli}
\alias{htmlize_uoli}
\title{Creat an unordered list (with bullet points) of texts.}
\usage{
htmlize_uoli(...)
}
\arguments{
\item{...}{One or more strings.}
}
\value{
HTML code wrapped around the input text.
}
\description{
Feed this function however many lines of strings you want, and it'll itemize
them into a bullet-pointed list.
}
\examples{
s1 = 'Hi, my name is GM. How are you?'
s2 = "I'm fine. Thank you, and yourself?"
s3 = "I'm doing super. Nice meeting you."
htmlize_uoli(s1, s2, s3)
}
