% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlize.R
\name{htmlize_w_linebreaks}
\alias{htmlize_w_linebreaks}
\title{Insert line breaks between text.}
\usage{
htmlize_w_linebreaks(...)
}
\arguments{
\item{...}{One or more strings.}
}
\value{
HTML code wrapped around the input text.
}
\description{
Feed this function however many lines of strings you want, and it'll insert
a line break between every two consecutive lines.
}
\examples{
s1 = 'Hi, my name is GM. How are you?'
s2 = "I'm fine. Thank you, and yourself?"
s3 = "I'm doing super. Nice meeting you."
htmlize_w_linebreaks(s1, s2, s3)
}
