% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intbox-n-ddboxes.R
\name{intbox_n_ddboxes_server}
\alias{intbox_n_ddboxes_server}
\title{Implement 1 integer input box and 2 dropdown boxes together.}
\usage{
intbox_n_ddboxes_server(
  id,
  labels = c("x-axis", "y-axis"),
  choice_part_labels = "X",
  selections = c("X1", "X2")
)
}
\arguments{
\item{id}{String. The input slot that will be used to access the value.}

\item{labels}{Character vector of display labels for the dropdown boxes.}

\item{choice_part_labels}{String. Part of the label for the value choices.}

\item{selections}{Character vector of the value choices for each dropdown box.
All dropdown boxes share the same value choices.}
}
\value{
A module server function that can be called in a reactive env.
}
\description{
The server component of the Shiny module for creating a integer input box and
2 side-by-side dropdown boxes with choices depending on the value entered
into the input box.
}
\examples{
inst/examples/ex-intbox_n_ddboxes.R
}
\seealso{
\code{\link{intbox_n_ddboxes_ui}} for the UI.
}
