% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intbox-n-ddboxes.R
\name{intbox_n_ddboxes_ui}
\alias{intbox_n_ddboxes_ui}
\title{A group of 1 integer input box and 2 dropdown boxes.}
\usage{
intbox_n_ddboxes_ui('id', 'label', min = 2, max = 10000000, value = 2)
}
\arguments{
\item{id}{String. The input slot that will be used to access the value.}

\item{label}{String (default is NULL). Display label for the input box.}

\item{min}{Integer (default = 2). Minimal value it takes.}

\item{max}{Integer (default = 10000000). Maximum value it takes.}

\item{value}{Integer (default is the same as min). Initial value it takes.}
}
\value{
Must use it together with its counterpart server.
}
\description{
The UI component of the Shiny module for creating one integer input box and
below it 2 side-by-side dropdown boxes with choices depending on the value
entered into the input box.
}
\examples{
inst/examples/ex-intbox_n_ddboxes.R
}
\seealso{
\code{\link{intbox_n_ddboxes_server}} for the server.
}
