% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletmap_module.R
\name{leafletMap}
\alias{leafletMap}
\title{Leaflet map module, Server logic}
\usage{
leafletMap(
  input,
  output,
  session,
  data,
  date_col = "datetime",
  points_col = "person_id",
  v_anim = 24,
  traj_col = "trajectory_id"
)
}
\arguments{
\item{data}{A Dataframe, is supposed to contains "lat" and "lng" cols, 
otherwise things won't work}

\item{date_col}{Chr. Name of the column identifying Date time reference.}

\item{points_col}{Chr.Name of the colum of the Points in the map. (circles)
The trajectories(polylines) are always represented with respect to this column reference.}

\item{v_anim}{num. Speed of the time animation. \code{1} means 5 hours.}

\item{traj_col}{Column idenifying the trajectories}
}
\value{
A shiny Module
}
\description{
Leaflet map module, Server logic
}
\examples{

require(shiny)
require(dplyr)
if(interactive()) {
df <- feather::read_feather("C://Users//Utente//Desktop//Maputo//transport_mode_detection//data_speed_acc_new.feather") \%>\%
  slice(1:100)
df <- rename(df, time = datetime, pid  = person_id, tid  = trajectory_id,)
ui <- fluidPage(
  leafletMapUi("map")
)
server <- function(input, output, session) {
  callModule(leafletMap, "map", data = df, v_anim = 1, date_col = "time", points_col = "pid", traj_col = "tid")
}
shinyApp(ui, server)
}
}
