% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletmap_module.R
\name{leafletMapUi}
\alias{leafletMapUi}
\title{Ui module for Shiny interactive map}
\usage{
leafletMapUi(id)
}
\arguments{
\item{id}{shiny modue identification}
}
\value{
Server function
}
\description{
@seealso \link{leafletMap}
}
