% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlize.R
\name{linebreaks}
\alias{linebreaks}
\title{Multiple Linebreaks}
\usage{
linebreaks(n)
}
\arguments{
\item{n}{Integer. The number of line breaks to make.}
}
\value{
HTML code of n linebreaks.
}
\description{
Creates n linebreaks.
}
\examples{
linebreaks(5)
}
