% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata_module.R
\name{loadfile}
\alias{loadfile}
\title{Load Data in Shiny. Server function}
\usage{
loadfile(input, output, session)
}
\value{
A reactive Dataframe.
}
\description{
Choose between .txt, .csv, .FEATHER format.
}
\examples{
require(shiny)
if (interactive()) {
ui <- 
  fluidPage(
    sidebarLayout(
      sidebarPanel(
        loadfileUI(id = "read")
    ), 
      mainPanel(tableOutput("tbl"))
  )
)


server <- function(input, output, session) {
  
  data <- callModule(loadfile, "read")
  
  output$tbl <- renderTable({
    validate(need(data(), message = FALSE))
    
    head(data()())
  })
}

shinyApp(ui, server)
}
}
