% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata_module.R
\name{loadfileUI}
\alias{loadfileUI}
\title{Load Data UI module}
\usage{
loadfileUI(id)
}
\arguments{
\item{id}{UI identification}
}
\value{
Ui function.
}
\description{
Load Data in shiny. Choose between .txt, .csv, .FEATHER format
}
