% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logi_test}
\alias{logi_test}
\title{Generalise logical statements
look if x is equal to the dots arguments. If ... is a list, then you can specify
a function rule in type, passing by name.}
\usage{
logi_test(x, ..., type = "any")
}
\arguments{
\item{x}{A character}

\item{...}{A vector of character}

\item{type}{a function}
}
\value{
A logical
}
\description{
Generalise logical statements
look if x is equal to the dots arguments. If ... is a list, then you can specify
a function rule in type, passing by name.
}
\examples{
logi_test("ciao", c("come", "va", "ciao"))
          logi_test("ciao", c("come", "va", "ciao"), type = "all")
}
