% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{look_for}
\alias{look_for}
\title{look for a specific element inside a list}
\usage{
look_for(x, look = "transformed")
}
\arguments{
\item{x}{list}

\item{look}{name of elemet in the list to look for}
}
\value{
list
}
\description{
look for a specific element inside a list
}
