% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkboxes-allon-alloff.R
\name{mk_checkboxes}
\alias{mk_checkboxes}
\title{Grouped checkboxes.}
\usage{
mk_checkboxes(id, label = NULL, choices, selected = NULL)
}
\arguments{
\item{id}{String. The input slot that will be used to access the value.}

\item{label}{String. String. Display label/title for the checkboxes group, or
NULL (default) for no label.}

\item{choices}{Character vector. List of values to show checkboxes for.
If elements of the list are named then that name rather than the value
is displayed to the user. The values should be strings; other types
(such as logicals and numbers) will be coerced to strings.}

\item{selected}{The values that should be initially selected, if any.
Default is NULL for no initial selection.}
}
\value{
A list of HTML elements that can be added to a UI definition.
}
\description{
Create a group of checkboxes that can be used to toggle multiple choices
independently. The server will receive the input as a character vector of the
selected values.
}
\seealso{
\code{\link{checkbox_w_onoff_button_server}} for an example of how
         to use this function.
}
