% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk-dropdown-box.R
\name{mk_dropdown_box}
\alias{mk_dropdown_box}
\title{A plain dropdown box.}
\usage{
mk_dropdown_box(id, label = NULL, choices)
}
\arguments{
\item{id}{String. The input slot that will be used to access the value.}

\item{label}{String. Display label/title for the dropdown box, or
NULL (default) for no label.}

\item{choices}{Character vector. List of values listed under the dropdown box
for user to select. If elements of the list are named then that name
rather than the value is displayed. The values should be strings;
other types (such as logicals and numbers) will be coerced to strings.}
}
\value{
A list of HTML elements that can be added to a UI definition.
}
\description{
Create a plain dropdown box where only one value can be selected.
}
\examples{
inst/examples/ex-dropdown-box.R
}
\seealso{
\code{\link{dropdown_box_ui}} for an example of how to use it.
}
