% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_run_button.R
\name{mk_run_button}
\alias{mk_run_button}
\title{Run button.}
\usage{
mk_run_button(id, label = "Run")
}
\arguments{
\item{id}{String. The input slot that will be used to access the value.}

\item{label}{String. Text (default = 'Run') to be shown on the run button.}
}
\value{
A list of HTML elements that can be added to a UI definition.
}
\description{
Create a run button that triggers some calculation when clicked. Run buttons
are often used in `observeEvent()`.
}
