% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk-slider.R
\name{mk_slider}
\alias{mk_slider}
\title{An awesome slider.}
\usage{
mk_slider(id, label = "Slider Input", ...)
}
\arguments{
\item{id}{String. The input slot that will be used to access the value.}

\item{label}{String. Text label of the slider (default = 'Slider Input').}
}
\value{
A list of HTML elements that can be added to a UI definition.
}
\description{
Create a pretty slider that takes input from the user as they slide it.
}
\examples{
inst/examples/ex-mk-slider.R
}
