% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functioncalls.R
\name{modify_call}
\alias{modify_call}
\title{modify the function call x with a list of args}
\usage{
modify_call(x, args)
}
\arguments{
\item{x}{A function Call}

\item{args}{list of Arguments, the names correspond to the names of the formals of the function.}
}
\value{
A function call
}
\description{
modify the function call x with a list of args
}
