% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatransformation.R
\name{my_boxcox}
\alias{my_boxcox}
\title{BoxCox transformation of x}
\usage{
my_boxcox(x)
}
\arguments{
\item{x}{num vector}
}
\value{
list. transformed x, lambda optimal, formula if \code{x} has negative values
}
\description{
BoxCox transformation of x
}
