% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maptools.R
\name{my_leafletproxy}
\alias{my_leafletproxy}
\title{Add polylines to an existing map}
\usage{
my_leafletproxy(data, grp_f, clr = "blue", proxy, label, criteria, traj_col)
}
\arguments{
\item{data}{Data containing lynes, and lng, lat cols}

\item{grp_f}{Levels of the group accepted. The group is identified by \code{criteria}}

\item{clr}{chr. Identy the colors of the lines}

\item{proxy}{Existing Leaflet map.}

\item{label}{Chr. Label pf the line}

\item{criteria}{Chr. Name of the column identifying the filter group.}

\item{traj_col}{Chr. Name of the column identifying the Trajectories}
}
\value{
A Leaflet map
}
\description{
Add polylines to an existing map
}
