% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{normal_plots}
\alias{normal_plots}
\title{Graphical representation of Normal assumption, Univariate data}
\usage{
normal_plots(
  x,
  title_list = list(hist = "histogram", scatter = "scatter", qqnorm =
    "Theoretical vs Sample quantiles")
)
}
\arguments{
\item{x}{Num Vector}

\item{title_list}{Optional}
}
\value{
list of plots: Scatter of \code{x}, histogram of \code{x}, qqplot of \code{x}
}
\description{
Graphical representation of Normal assumption, Univariate data
}
