% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_module.R
\name{pca}
\alias{pca}
\title{Perform PCA analysis on shiny}
\usage{
pca(input, output, session, data)
}
\arguments{
\item{data}{A Dataframe}
}
\value{

}
\description{
Perform PCA analysis on shiny
}
\examples{
require(shiny)
if(interactive()) {

ui <- fluidPage(
  title = "Principal Component Analysis",
  pcaUI("pca")

  )

server <- function(input, output, session) {

  callModule(pca, "pca", data = mtcars)
}
shinyApp(ui, server)
} 
}
