% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radio-buttons.R
\name{radio_buttons_ui}
\alias{radio_buttons_ui}
\title{A group of pretty radio buttons.}
\usage{
radio_buttons_ui(id, label = NULL, choices)
}
\arguments{
\item{id}{String. The input slot that will be used to access the value.}

\item{label}{String. Display label/title for the radio buttons group, or
NULL (default) for no label.}

\item{choices}{Character vector. List of values to show radio buttons for.
If elements of the list are named then names rather than values are
displayed. The values should be strings; other types
(such as logicals and numbers) will be coerced to strings.}
}
\value{
Must use it together with its counterpart server.
}
\description{
The UI component of the Shiny module for creating a group of pretty radio
buttons for selecting one value among a list of values.
}
\examples{
inst/examples/ex-radio-buttons.R
}
\seealso{
\code{\link{radio_buttons_server}} for the server
         \code{\link{mk_radio_buttons}} for how to make a group of awesome
         radio buttons.
}
