% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata_module.R
\name{txtFile}
\alias{txtFile}
\title{Server function for loading txt files}
\usage{
txtFile(input, output, session, stringsAsFactors = FALSE)
}
\arguments{
\item{stringsAsFactors}{FALSE by default}
}
\value{
A A reactiv dataframe
}
\description{
Server function for loading txt files
}
