% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_module.R
\name{typeplot}
\alias{typeplot}
\title{Server function of Data Visualization in Shiny}
\usage{
typeplot(input, output, session, data)
}
\arguments{
\item{data}{A Dataframe}
}
\value{
A Shiny server module
}
\description{
Server function of Data Visualization in Shiny
}
\examples{
# an example 
require(shiny)
if(interactive()) {

ui <- fluidPage(

  typeplotUi("type")


)

server <- function(input, output, session) {

  callModule(typeplot, "type", data = CO2)

}


shinyApp(ui, server)

}
}
