% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_module.R
\name{typeplotUi}
\alias{typeplotUi}
\title{Automatic Data Visualization in Shiny}
\usage{
typeplotUi(id)
}
\arguments{
\item{id}{Identification Module}
}
\value{
Ui module
}
\description{
Automatic Data Visualization in Shiny
}
