% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{upload_server}
\alias{upload_server}
\title{Implement the upload button.}
\usage{
upload_server(id)
}
\arguments{
\item{id}{String. The input slot that will be used to access the value.}
}
\value{
A module server function that can be called in a reactive env to
        read the user-uploaded file and return a data frame.
}
\description{
The server component of the Shiny module for creating a pretty upload button
for uploading a text file.
}
\examples{
inst/examples/ex-upload.R
}
\seealso{
\code{\link{upload_ui}} for the UI.
}
