% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{upload_ui}
\alias{upload_ui}
\title{Upload button.}
\usage{
upload_ui(id, label = "Select a text file (< 30MB)")
}
\arguments{
\item{id}{String. The input slot that will be used to access the value.}

\item{label}{String. Text (default = "Select a text file (< 30MB)") to be
shown on the upload button.}
}
\value{
Must use it together with its counterpart server.
}
\description{
The UI component of the Shiny module for creating a pretty upload button
for uploading a text file.
}
\examples{
inst/examples/ex-upload.R
}
\seealso{
\code{\link{upload_server}} for the server.
}
