% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{user_filter}
\alias{user_filter}
\title{Filter chr or num vars}
\usage{
user_filter(x, param, val)
}
\arguments{
\item{x}{df}

\item{param}{col name}

\item{val}{numeric interval or charachter string}
}
\value{
filtered df
}
\description{
Filter chr or num vars
}
