% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{user_filter_m}
\alias{user_filter_m}
\title{Simultaneous filter}
\usage{
user_filter_m(x, params, vals)
}
\arguments{
\item{x}{df}

\item{params}{list of params}

\item{vals}{list of vals}
}
\value{
filtered df
}
\description{
Simultaneous filter
}
