% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maptools.R
\name{wrap_myleaflet}
\alias{wrap_myleaflet}
\title{Wrap of my leaflet. The Arguments are directly passed in \link{my_leafletproxy}
This is used to distinct different "types" of Trajectories}
\usage{
wrap_myleaflet(data, gruppo_list, color_list, proxy, label, criteria, traj_col)
}
\arguments{
\item{gruppo_list}{chr or list. Passed to map}

\item{color_list}{chr or list. Passed to map}
}
\value{
list of updated leaflet maps
}
\description{
Wrap of my leaflet. The Arguments are directly passed in \link{my_leafletproxy}
This is used to distinct different "types" of Trajectories
}
