% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkboxes-allon-alloff.R
\name{checkboxes_allon_alloff_ui}
\alias{checkboxes_allon_alloff_ui}
\title{A group of checkboxes with all-on/all-off control buttons.}
\usage{
checkboxes_allon_alloff_ui(id, bttn_height = 60)
}
\arguments{
\item{id}{String. The input slot that will be used to access the value.}

\item{bttn_height}{Integer. It controls the space below the all-on/all-off
buttons. The bigger the value (default = 60), the more space there is.
Without it (or when its value is too small), it can happen that the
all-on and all-off buttons will overlap with whatever UI elements
that are placed right below them.}
}
\value{
A list of HTML elements that can be added to a UI definition and
        accessed by the server component.
}
\description{
The UI component of the Shiny module for creating a group of pretty
checkboxes with an all-on and an all-off control buttons.
}
\examples{
inst/examples/ex-checkboxes-allon-alloff.R
}
\seealso{
\code{\link{checkboxes_allon_alloff_server}} for the server.
}
