% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_ints.R
\name{is_integer}
\alias{is_integer}
\title{Check if a value is an integer or not.}
\usage{
is_integer(x)
}
\arguments{
\item{x.}{A value of any type.}
}
\value{
Returns TRUE or FALSE.
}
\description{
Base R lacks a function for checking if a value is an integer or not. This
function does that.
}
\examples{
is_integer(NA_integer_)
is_integer(NA_complex_)
is_integer(NA_real_)
is_integer(NA_character_)
is_integer(-9)
is_integer(9)
is_integer(-9.2)
is_integer(12)
is_integer(NaN)
is_integer(Inf)
is_integer(-Inf)
}
