% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_dend.R
\name{plt_dend}
\alias{plt_dend}
\title{Plot dendrogram.}
\usage{
plt_dend(
  hc,
  n_clusters,
  label_size = 0.8,
  title = "Hiearchical Clustering Result",
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{hc}{Object returned by `hclust()`.}

\item{n_clusters}{Integer. Number of clusters.}

\item{label_size}{Numeric. Size (default = 0.8) of text labels for the leaves.}

\item{title}{String. Plot title (default = "Hiearchical Clustering Result").}

\item{subtitle}{String. Plot subtitle (default = NULL).}

\item{xlab}{String. Label of the x-axis (default = NULL).}

\item{ylab}{String. Label of the y-axis (default = NULL).}
}
\value{
A ggplot2 dendrogram.
}
\description{
Visualize hierarchical clustered object with different colored clusters and
leaf labels.
}
\examples{
data(USArrests)
dd = dist(scale(USArrests), method = "euclidean")
hc = hclust(dd, method = "ward.D2")
plt_dend(hc, 2)
plt_dend(hc, 4, label_size = 0.5)
plt_dend(hc, 5)
}
