% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Scatter plot}
\usage{
scatter_plot(
  df,
  x = NULL,
  y,
  x_limits = NULL,
  y_limits = NULL,
  color = NULL,
  title = "Scatter",
  wrap = NULL
)
}
\arguments{
\item{df}{dataframe}

\item{x}{col name of x axis}

\item{y}{col name of y axis}

\item{x_limits}{interval to filter x axis}

\item{y_limits}{interval to filter y axis}

\item{color}{chr.  Name of the variable to which apply color logic.}

\item{title}{Optional.}

\item{wrap}{chr. column to which apply facet_wrap, y scale is free}
}
\value{
ggplot object
}
\description{
Scatter plot
}
